major restructure to clean up components
