+++ components/Lobby/Index.tsx
@@ -3,28 +3,14 @@ import React, { useEffect, useState, useRef } from "react";
 import styled from "styled-components";
 import { AnimatePresence, motion } from "framer-motion";
 import Chat from "../Chat";
-import { convertToNum } from "../../utils/users";
 import { variables } from "@/styles/Variables";
-import { buttonType, h2styles, h3styles, pLarge, pSmall } from "@/styles/Type";
-import PlayersInLobby from "./PlayersInLobby/Index";
+import { buttonType, h2styles, pLarge, pSmall } from "@/styles/Type";
 import { MediaQueries } from "@/styles/Utilities";
+import FullGame from "../Game/FullGame/Index";
+import LobbyInfo from "./LobbyInfo/Index";
 
 const Player = styled.div`
   ${pSmall}
-`;
-const PlayerContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-`;
-
-const CloverIcon = styled.img`
-  width: 50px;
-  height: 50px;
-  padding: 6px;
-  border-radius: 50%;
-  border: 1px solid white;
 `;
 
 const MainContainer = styled.div`
@@ -44,59 +30,9 @@ const MainInnerContainer = styled.div`
   grid-template-columns: 360px 880fr;
 `;
 
-const CardContainer = styled.div`
+const FullGameContainer = styled.div`
   display: flex;
   flex-direction: column;
-`;
-
-const IndividualCard = styled.div`
-  min-width: 50px;
-`;
-
-const ImageOfCard = styled(motion.img)`
-  width: 50px;
-`;
-
-const IndividualCardContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-`;
-
-const MainButtonsContainer = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const BtnContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 24px;
-  /* position: absolute; */
-  bottom: 0;
-`;
-
-const CardsContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-`;
-
-const Message = styled.div`
-  display: flex;
-  flex-direction: column;
-  /* max-width: 500px; */
-  width: 50%;
-  height: 50%;
-  align-items: center;
-  justify-content: center;
-`;
-
-const PlayerAndCardContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 50px;
-  margin-bottom: 30px;
 `;
 
 const Header = styled.h2`
@@ -108,43 +44,6 @@ const Header = styled.h2`
 
 const Button = styled.button`
   ${buttonType}
-`;
-
-const GameButtonContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-`;
-
-const GameContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-`;
-
-const CurrentPlayerCard = styled(IndividualCard)`
-  z-index: 1;
-`;
-
-const OtherPlayersCard = styled(IndividualCard)`
-  opacity: 0.5;
-`;
-
-const PlayerName = styled.p`
-  ${pSmall}
-`;
-
-const PlayerUpNext = styled.p`
-  ${pSmall}
-  margin: 30px 0 20px;
-`;
-
-const LobbyInfo = styled.div`
-  display: flex;
-  flex-direction: column;
 `;
 
 const CorrectMessaging = styled(motion.div)`
@@ -206,42 +105,6 @@ const IncorrectMessaging = styled(motion.div)`
   }
 `;
 
-const CloseModal = styled.button`
-  background-color: unset;
-  border: unset;
-  z-index: 100;
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  cursor: pointer;
-  width: 34px;
-  height: 34px;
-
-  @media ${MediaQueries.mobile} {
-    width: 44px;
-    height: 44px;
-    padding: 10px;
-    right: 8px;
-    top: 8px;
-  }
-
-  svg {
-    polyline {
-      stroke: #fff;
-      transition: filter ease-out 0.35s;
-    }
-  }
-
-  &:hover,
-  &:focus {
-    svg {
-      polyline {
-        filter: brightness(0.7);
-      }
-    }
-  }
-`;
-
 const Description = styled.p`
   ${pSmall}
 `;
@@ -256,7 +119,6 @@ interface GameData {
   cardData: [];
 }
 
-// Type for a Signle Card that the Card Game API give me
 type SingleCard = {
   player: string;
   image: string;
@@ -269,7 +131,6 @@ type SingleCard = {
   cardNext?: boolean;
 };
 
-// All the cards in the Player array
 type Player = SingleCard[];
 type PlayerData = {};
 
@@ -282,7 +143,6 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
 
   const [usersLockedIn, setUsersLockedIn] = useState(false);
   const [gameStarted, setGameStarted] = useState(false);
-  // Initialize state for the current round
   const [currentRound, setCurrentRound] = useState(0);
 
   const [booleanMessage, setBooleanMessage] = useState(null);
@@ -300,10 +160,6 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
 
   const otherPlayers = users?.filter(
     (user: any) => user?.username !== username
-  );
-
-  const currentPlayer = users?.filter(
-    (user: any) => user?.username === username
   );
 
   const lockInPlayersHandler = () => {
@@ -325,164 +181,6 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
       currentRound,
     });
     setGameStarted(true);
-  };
-
-  const gameLogicHandler = (option: string) => {
-    const currentPlayerCard: SingleCard | undefined =
-      allGameData?.cardData[currentPlayerIndex][currentRound];
-
-    if (!currentPlayerCard) {
-      // Handle the case where currentPlayerCard is undefined
-      return;
-    }
-
-    const updatedCard: SingleCard = {
-      ...((currentPlayerCard as SingleCard) || {}),
-      selectedOption: option,
-    };
-
-    // Update the cardNext for the next player
-    const nextPlayerIndex = (currentPlayerIndex + 1) % users.length;
-    const nextPlayerCard: SingleCard | any =
-      allGameData?.cardData[nextPlayerIndex][currentRound];
-
-    const nextUpdatedCard: SingleCard | any = {
-      ...nextPlayerCard,
-      cardNext: true,
-    };
-
-    const updatedPlayerData: PlayerData = (allGameData?.cardData || []).map(
-      (player: Player, index) =>
-        index === currentPlayerIndex
-          ? player.map((card) =>
-              card === currentPlayerCard ? updatedCard : card
-            )
-          : index === nextPlayerIndex
-          ? player.map((card) =>
-              card === nextPlayerCard ? nextUpdatedCard : card
-            )
-          : player
-    );
-
-    const card: Player | any =
-      allGameData?.cardData[currentPlayerIndex][currentRound];
-    const prevCard = allGameData?.cardData[currentPlayerIndex][0];
-
-    const prevPrevCard = allGameData?.cardData[currentPlayerIndex][1];
-
-    const { suit, value }: any = card || {};
-    const { value: prevValue }: any = prevCard || {};
-    const { value: prevPrevValue }: any = prevPrevCard || {};
-
-    let isCorrect = false;
-
-    switch (option) {
-      case "red":
-        isCorrect = suit === "HEARTS" || suit === "DIAMONDS";
-        break;
-
-      case "black":
-        isCorrect = suit === "SPADES" || suit === "CLUBS";
-        break;
-
-      case "lower":
-        isCorrect = convertToNum(prevValue) > convertToNum(value);
-        break;
-
-      case "higher":
-        isCorrect = convertToNum(prevValue) < convertToNum(value);
-        break;
-
-      case "in":
-        isCorrect =
-          // prevCard == 3
-          // prevPrevCard = 10
-          (convertToNum(prevValue) < convertToNum(value) &&
-            convertToNum(prevPrevValue) > convertToNum(value)) ||
-          (convertToNum(prevValue) > convertToNum(value) &&
-            convertToNum(prevPrevValue) < convertToNum(value));
-        break;
-
-      case "out":
-        isCorrect =
-          (convertToNum(prevValue) > convertToNum(value) &&
-            convertToNum(prevPrevValue) > convertToNum(value)) ||
-          (convertToNum(prevValue) < convertToNum(value) &&
-            convertToNum(prevPrevValue) < convertToNum(value));
-        break;
-
-      case "spade":
-        isCorrect = suit === "SPADES";
-        break;
-
-      case "diamond":
-        isCorrect = suit === "DIAMONDS";
-        break;
-
-      case "heart":
-        isCorrect = suit === "HEARTS";
-        break;
-
-      case "club":
-        isCorrect = suit === "CLUBS";
-        break;
-
-      default:
-        break;
-    }
-
-    //TODO: need to add back some logic about the # equaling eachother
-
-    let selectionMessage = isCorrect ? true : false;
-
-    socket.emit("send_modal_active", {
-      roomId,
-      activeModal: true,
-    });
-
-    socket.emit("send_current_player_message", {
-      roomId,
-      selectionMessage,
-      currentUsersMessageTrue: `
-      CORRECT!
-      ${
-        card?.player
-      } guessed ${option}! The card was a: ${card?.value.toLowerCase()} of ${card?.suit.toLowerCase()}`,
-      currentUsersMessageFalse: `${isCurrentPlayer} was incorrect and is drinking!`,
-    });
-
-    socket.emit("send_other_players_message", {
-      roomId,
-      selectionMessage,
-      otherUsersMessageTrue: `One moment...${isCurrentPlayer} is choosing who drinks`,
-      otherUsersMessageFalse: `${isCurrentPlayer} was incorrect and is drinking!`,
-      currentUsersMessageTrue: `
-      CORRECT!
-      ${
-        card?.player
-      } guessed ${option}! The card was a: ${card?.value.toLowerCase()} of ${card?.suit.toLowerCase()}`,
-    });
-
-    socket.emit("updated_card_data", {
-      users,
-      roomId,
-      cardData: updatedPlayerData,
-    });
-
-    socket.emit("current_index", {
-      users,
-      roomId,
-      cardData: updatedPlayerData,
-      currentPlayerIndex,
-    });
-
-    socket.emit("send_current_player", {
-      roomId,
-      isCurrentPlayer,
-      currentRound,
-    });
-
-    // setActiveModal(true);
   };
 
   let usersToDrinkArr: any = [];
@@ -599,8 +297,6 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
     countdown,
   ]);
 
-  const ref = useRef(null);
-
   const motionProps = {
     initial: {
       opacity: 0,
@@ -619,11 +315,6 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
     },
   };
 
-  const hideModal = () => {
-    setActiveModal(false);
-    setBooleanMessage(null);
-  };
-
   return (
     <MainContainer>
       <MainInnerContainer>
@@ -633,169 +324,27 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
           roomId={roomId}
           users={users}
         />
-        <CardContainer>
+        <FullGameContainer>
           <Header>
             Lobby for Room id: <b>{roomId}</b>
           </Header>
-          <LobbyInfo>
-            {!gameStarted && <PlayersInLobby users={users} />}
-            {!gameStarted && (
-              <GameButtonContainer>
-                {users.length > 0 && !usersLockedIn && !gameStarted ? (
-                  <Button onClick={lockInPlayersHandler}>
-                    Lock in players
-                  </Button>
-                ) : (
-                  <Button onClick={startGameHandler}>Start Game</Button>
-                )}
-              </GameButtonContainer>
-            )}
-          </LobbyInfo>
-          <GameContainer>
-            {allGameData &&
-              allGameData?.cardData.map(
-                (player: Player, playerIndex: number) => {
-                  const CardComponent =
-                    currentPlayerIndex === playerIndex
-                      ? CurrentPlayerCard
-                      : OtherPlayersCard;
-
-                  return (
-                    <PlayerAndCardContainer>
-                      <PlayerContainer className="item">
-                        <CloverIcon src="clover.svg" alt="clover" />
-                        <Player>{player[playerIndex].player} </Player>
-                      </PlayerContainer>
-                      {/* <PlayerName>{player[playerIndex].player}</PlayerName> */}
-                      <CardsContainer ref={ref}>
-                        {player.map((singleCard: SingleCard, index: number) => {
-                          return (
-                            <>
-                              <IndividualCardContainer key={`player-${index}`}>
-                                {singleCard.selectedOption ? (
-                                  <ImageOfCard
-                                    key={`${singleCard.selectedOption}-${singleCard.image}`}
-                                    initial={{
-                                      opacity: 0,
-                                      rotateX: 360,
-                                      rotateY: 720,
-                                      scale: 0,
-                                    }}
-                                    animate={{
-                                      rotateX: 0,
-                                      opacity: 1,
-                                      rotateY: 0,
-                                      scale: 1,
-                                    }}
-                                    exit={{
-                                      opacity: 0,
-                                      rotateX: 360,
-                                      rotateY: 720,
-                                      scale: 0,
-                                    }}
-                                    transition={{
-                                      duration: `0.8`,
-                                      ease: "easeInOut",
-                                    }}
-                                    src={singleCard.image}
-                                  />
-                                ) : (
-                                  // <p>{singleCard.code}</p>
-                                  <ImageOfCard
-                                    key={`default-${singleCard.code}`}
-                                    initial={{
-                                      opacity: 0,
-                                      rotateX: 360,
-                                      rotateY: 720,
-                                      scale: 0,
-                                    }}
-                                    animate={{
-                                      rotateX: 0,
-                                      opacity: 1,
-                                      rotateY: 0,
-                                      scale: 1,
-                                    }}
-                                    exit={{
-                                      opacity: 0,
-                                      rotateX: 360,
-                                      rotateY: 720,
-                                      scale: 0,
-                                    }}
-                                    transition={{
-                                      duration: `0.5`,
-                                      ease: "easeInOut",
-                                    }}
-                                    src="green_card.png"
-                                  />
-                                )}
-                              </IndividualCardContainer>
-                            </>
-                          );
-                        })}
-                      </CardsContainer>
-                    </PlayerAndCardContainer>
-                  );
-                }
-              )}
-            {allGameData && (
-              <PlayerUpNext>Player up next: {isCurrentPlayer}</PlayerUpNext>
-            )}
-            {allGameData && (
-              <MainButtonsContainer>
-                {users[currentPlayerIndex]?.username === username &&
-                  currentRound === 0 && (
-                    <BtnContainer className="btn-container">
-                      <Button onClick={() => gameLogicHandler("red")}>
-                        Red
-                      </Button>
-                      <Button onClick={() => gameLogicHandler("black")}>
-                        Black
-                      </Button>
-                    </BtnContainer>
-                  )}
-                {users[currentPlayerIndex]?.username === username &&
-                  currentRound === 1 && (
-                    <BtnContainer className="btn-container">
-                      <Button onClick={() => gameLogicHandler("lower")}>
-                        Lower
-                      </Button>
-                      <Button onClick={() => gameLogicHandler("higher")}>
-                        Higher
-                      </Button>
-                    </BtnContainer>
-                  )}
-                {users[currentPlayerIndex]?.username === username &&
-                  currentRound === 2 && (
-                    <BtnContainer className="btn-container">
-                      <Button onClick={() => gameLogicHandler("in")}>In</Button>
-                      <Button onClick={() => gameLogicHandler("out")}>
-                        Out
-                      </Button>
-                    </BtnContainer>
-                  )}
-                {users[currentPlayerIndex]?.username === username &&
-                  currentRound === 3 && (
-                    <BtnContainer className="btn-container">
-                      <Button onClick={() => gameLogicHandler("club")}>
-                        Club
-                      </Button>
-                      <Button onClick={() => gameLogicHandler("spade")}>
-                        Spade
-                      </Button>
-                      <Button onClick={() => gameLogicHandler("diamond")}>
-                        Diamond
-                      </Button>
-                      <Button onClick={() => gameLogicHandler("heart")}>
-                        Heart
-                      </Button>
-                    </BtnContainer>
-                  )}
-              </MainButtonsContainer>
-            )}
-
-            {currentRound === 4 && <p>GAME IS OVER</p>}
-          </GameContainer>
-          {/* )} */}
+          <LobbyInfo
+            gameStarted={gameStarted}
+            users={users}
+            usersLockedIn={usersLockedIn}
+            startGameHandler={startGameHandler}
+            lockInPlayersHandler={lockInPlayersHandler}
+          />
+          <FullGame
+            username={username}
+            users={users}
+            allGameData={allGameData}
+            currentPlayerIndex={currentPlayerIndex}
+            currentRound={currentRound}
+            isCurrentPlayer={isCurrentPlayer}
+            socket={socket}
+            roomId={roomId}
+          />
           <AnimatePresence mode="wait">
             {booleanMessage !== null && activeModal && (
               <>
@@ -833,7 +382,7 @@ const GameLobby = ({ socket, username, roomId, users, showChat }: any) => {
               </>
             )}
           </AnimatePresence>
-        </CardContainer>
+        </FullGameContainer>
       </MainInnerContainer>
     </MainContainer>
   );
